+++ src/entities/post/api/createPost.ts
@@ -1,0 +1,23 @@
+import usePostsStore from "../../../features/posts/model/usePostsStore"
+
+// 게시물 추가를 위한 커스텀 훅
+export const useCreatePost = () => {
+  const { newPost, posts, setNewPost, setPosts, setShowAddDialog } = usePostsStore()
+
+  const addPost = async () => {
+    try {
+      const response = await fetch("/api/posts/add", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(newPost),
+      })
+      const data = await response.json()
+      setPosts([data, ...posts])
+      setShowAddDialog(false)
+      setNewPost({ title: "", body: "", userId: 1 })
+    } catch (error) {
+      console.error("게시물 추가 오류:", error)
+    }
+  }
+  return { addPost }
+}
