+++ src/features/posts/api/useUpdatePost.ts
@@ -1,0 +1,22 @@
+import usePostsStore from "../model/usePostsStore"
+
+// 게시물 업데이트
+export const useUpdatePost = () => {
+  const { selectedPost, posts, setPosts, setShowEditDialog } = usePostsStore()
+
+  const updatePost = async () => {
+    try {
+      const response = await fetch(`/api/posts/${selectedPost?.id}`, {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(selectedPost),
+      })
+      const data = await response.json()
+      setPosts(posts.map((post) => (post.id === data.id ? data : post)))
+      setShowEditDialog(false)
+    } catch (error) {
+      console.error("게시물 업데이트 오류:", error)
+    }
+  }
+  return { updatePost }
+}
